Rewrite tor-netdir::params to use tor-primitive-types
Rather than a hashmap from an enum to i32, instead we have an explicit type for each parameter.
The presence of a parameter is optional, allowing the user to decide how to handle its absence, e.g. .unwrap_or_default() to use the default value if missing.
Consequently, there are no longer runtime panics if fetching a particular value as the wrong type (get_u16, get_bool get).

Not yet implemented:
 * Tests
 * Parsing from string.