isolation: Add the mechanics for stream isolation
At this commit, the IsolationInfo structure is introduced which contains
IsolationFlag that dictate if a stream can be shared on a circuit with
others.

In order to have a commit that builds, the IsolationMap is removed from
crates/arti and thus empty isolation info objects are being used.

Next commits will add support for this in the arti SOCKS proxy
effectively refactoring the IsolationToken single use to a more complex
system of flags.

Part of #150

Signed-off-by: David Goulet <dgoulet@torproject.org>