arti: Isolate clients for the SOCKS proxy
The SOCKS proxy now keeps a map of TorClient and isolate them based on
an isolation key that is the source address, port and SOCKS
authentication if any.

At this commit, it is not possible yet to add isolation flag as they are
not yet available with the configuration file.

Closes #150

Signed-off-by: David Goulet <dgoulet@torproject.org>