Integration tests: Override CHUTNEY_PATH
If the user has their own CHUTNEY_PATH set (perhaps for use with
Tor), we want our integration tests to override that value so that
we use the freshly cloned Chutney instance.  Otherwise the network
will be created off in ${CHUTNEY_PATH}/net/nodes, and we won't
be able to find the arti.toml file.