tor-chanmgr: Drop FIXME about Channel, Arc and Clone
The comment says Arc is not needed because Channel is Clone.  That is
true.

But the code seems to clone the Arc<Channel> quite a lot, and Channel
is *nontrivial* to clone.  Indeed, it even directly contains an
Arc!  (As well as various cloneable interthread comms objeets.)  So
cloning the Channel is definitely more expensive than cloning an Arc
that would wrap it.  In the absence of benchmarking, I would guess
that this outweighs the extra indirection.

The fact that Arc<C> is open-coded everywhere at the call sites is
unfortunate.  It might be worth considering moving the Arc into
channel.rs.  Ie, deleting all the ad-hoc Arc<> and instead, have in
channel.rs's `type Channel = Arc<ChannelContents>`.