chanmgr: Remove Arcs around Channel
The Arc was used by the manager to allow it to clone the channel to
return it, but typically it would be about to be used in a circuit, in
which case the actual Channel would need to be cloned anyway.

Now, Channel is cheap to clone, so all these Arcs are redundant.

We remove them from the trait definitions so they end up going away
from the tests too.