chanmgr: Placate clippy by adopting horrible coding style
Calling `.clone()` on an Arc is rejected in our codebase by this
clippy lint: `#![warn(clippy::clone_on_ref_ptr)]`

I very strongly disagree with the coding style that this lint
enforces.  As can be seen from the previous commits in this MR, that
coding style was seriously in my way and I had to undo a lot of it.

IMO this lint, and the resulting idiom `Arc::clone(&thingum)`:

 * Exposes implementation details in an awkward way, requiring
   much mechanical change when restructuring changes the level
   at which Arcs appear (as in this MR).

 * Doesn't help when the type is a handle that implements Clone to
   produce multiple functionally equivalent references to the same
   underlying machinery.  This is extremely common - most of our
   synchronisation primitives like channels and watches, and most of
   the key API objects in the tor* crates, are like that.

 * Is disgusting syntactic vinegar.

However, having had my rant I guess I can live with it.  I'm new to
the project and oughtn't to be fighting over coding style.