tor-error: Introduce conversion_imports
We are going to want to write quite a few of these impls.

One way to do this woudl be a custom derive macro; that would allow us
to annotate the crate-specific error enum definition with the kinds
and flags.  But it would make special cases more awkward, and be quite
a bit of work.

So failing that, we want to reduce the boilerplate as much as we can,
and have standardised naming in these conversion methods.