Temporarily disable clippy::needless_borrow on nightly
This currently produces false positives:
  https://github.com/rust-lang/rust-clippy/issues/8367
  https://github.com/rust-lang/rust-clippy/issues/8380
  https://github.com/rust-lang/rust-clippy/pull/8355

There are many many places in our codebase where this comes up;
whacking them individually is too painful.  Even putting them in every
crate's clippy set is bad enough!

We can revert this when the fix is in upstream.  To this end, I am subscribed.