arti-config: Take &dyn AsRef<str>, &dyn AsRef<Path>
These call sites are far from performance-critical.  Using generics
like this can easily lead to code bloat (and slow compile times).

This seems like an excellent place to use trait object references,
which are quite ergonomic.  In each case the result is that we insist
on taking the borrowed form, and making our own copy.  This is quite
fine here.  It wouldn't be on a hot path.

We must change all the call sites to pass a reference.

The only downside to this paradigm is that right now the error message
if you try to pass a fat pointer directly, rather than reborrowing it,
isn't great.  I don't think this is a serious problem but I have just
filed an ehancement request for rustc
  https://github.com/rust-lang/rust/issues/93596
about it.