circmgr: Add Status information to FallbackSet.
This field will track when we can retry, and is now used to see
which fallbacks are usable.  (It is not yet set on success or
failure.)

The design here is a bit tricky; I've been through a few iterations
of it, and this is the one that seems like it will give us
reasonable lifetimes and behavior.