Add and use status reporting APIs for FallbackSet.
Internally to the circmgr crate, we use a FallbackMonitor object
with an API similar to the GuardStatusMonitor.  Externally, we have
reporting functions to note success and failure.

With this change, we now notice when fallbacks have failed or given
us useless information, and we mark them as unusable for a while
before we consider retrying them.  This should be sufficient to
solve the problem observed in #406, where some kinds of failures to
connect to (or download from) fallback directories resulted in an
obnoxious busy-loop.