fs-mistrust: Actually encode and decode OsString usernames
It turns out that the `toml` crate can't handle OsString, since
`toml` doesn't support serialize_newtype_variant, and the `serde`
crate tries to serialize OsString using that method.

Found by inspecting test coverage.

It is possible that this whole MR is a mistake and we should just forbid
non-UTF8 usernames in our toml?