Implement a wrapper to count queued data.
For flow control, we're going to use this to count the total number
of _semantic data_ bytes queued on a channel between the circuit
reactor and the stream.  But it will also be useful when we
implement DoS prevention and we want to track the amount of _memory
allocated_ on such a queue.